Refactor readBytes
